requirements.txt
flask 
flask_sqlalchemy 
flask_cors 
requests
python-dateutil
google-api-python-client 
google-auth 
google-auth-oauthlib